# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BATIQA AI Guest Assistant API
  version: 1.0.0
  description: |
    API backend untuk asisten tamu Hotel BATIQA Pekanbaru. Menghubungkan
    tamu (via QR token) dengan pipeline AI (Gemini + fallback rule-based),
    pembuatan tiket operasional, dan dasbor staf secara real-time (SSE).

    ## Autentikasi
    - **Token tamu (QR)**: header `X-Guest-Token` atau query `?t=<token>`.
      Token dihasilkan dari `GET /api/rooms/{room}/qr`, bertanda tangan HMAC-SHA256.
    - **Staf**: header `Authorization: Bearer <token>` dari `POST /api/staff/login`.

    Kontrak lengkap dapat dilihat di UI Swagger interaktif pada `/api/docs`.
servers:
  - url: http://localhost:8080
    description: Lokal
...
paths:
...
  /api/hotel-info:
    get:
      tags: [Content]
      summary: Informasi hotel (publik)
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Item info hotel
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/HotelInfo'
  /api/recommendations:
    get:
      tags: [Content]
      summary: Rekomendasi lokal Pekanbaru (publik)
      parameters:
        - name: category
          in: query
          schema:
            enum: [restaurant, cafe, tourism, shopping, atm, transportation]
        - name: max_price
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Rekomendasi
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recommendation'
...
components:
  securitySchemes:
    GuestToken:
      type: apiKey
      in: header
      name: X-Guest-Token
      description: Token QR bertanda HMAC dari /api/rooms/{room}/qr (bisa juga via query ?t=)
    StaffAuth:
      type: http
      scheme: bearer
      bearerFormat: Token sesi staf (dari /api/staff/login)
  schemas:
...
    HotelInfo:
      type: object
      properties:
        id: { type: integer }
        category: { type: string }
        title: { type: string }
        content: { type: string }
    Recommendation:
      type: object
      properties:
        name: { type: string }
        category:
          type: string
          enum: [restaurant, cafe, tourism, shopping, atm, transportation]
        price_min:
          type: integer
          nullable: true
        price_max:
          type: integer
          nullable: true
        distance_km:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
...